tax-rules/requirements.txt
fivetran-connector-sdk==2.2.1
requests==2.32.5
beautifulsoup4==4.14.2
lxml==6.0.2
python-dateutil==2.9.0.post0
google-generativeai==0.8.3
python-dotenv==1.0.1
